FormRenderingTraitを追加し、cfFormMailerクラスのフォームレンダリング機能を強化
